refactor(store): split monolithic mcReducer into domain reducers
- Extracts token economy logic into `economyReducer.ts`
- Extracts mission and task scheduling logic into `missionReducer.ts`
- Extracts privilege logic into `privilegeReducer.ts`
- Extracts responsibility tracking into `responsibilityReducer.ts`
- Extracts global system settings and logs into `systemReducer.ts`
- Updates `mcReducer.ts` to act as the root delegator

This fulfills the AI-First Architecture guidelines by breaking down a nearly 600-line "God file" to improve context window efficiency and adhere to Single Responsibility Principle. All unit tests pass, and behavior is 100% equivalent to previous state.
