+++ .jules/architect-journal.md
@@ -30,3 +30,4 @@
 **Anti-Pattern: Hardcoded Colors Bypassing Design System.** The game introduces a standalone `snake.css` with raw hex values instead of utilizing Tailwind or `gcal-simplified` design tokens, violating the styling invariants.
 **Anti-Pattern: Unused Variables Breaking CI.** Leaving variables like `_livesRemaining` defined but unused in `QuizOverlay.tsx` triggers `--max-warnings 0` ESLint failures. Strict lint hygiene must be maintained before merging.
 **Regression Risk: Missing Coverage.** The core game engine (`useSnakeGame.ts`) and React overlays were committed without corresponding Vitest unit tests, while only `additionQuiz.test.ts` was implemented.
+## 2026-05-03 - Split mcReducer\n\n**Learning:** When refactoring large monolithic reducers (e.g., `mcReducer.ts`), separate them into smaller, domain-specific reducers (e.g., `economyReducer.ts`, `missionReducer.ts`) in a localized `reducers/` directory and compose them in the main file. Use existing split test files as a guide for identifying domain boundaries.\n**Action:** Always look for natural boundaries, such as those established by test files, to guide the separation of concerns when splitting large files.
