+++ src/mission-control/store/reducers/privilegeReducer.ts
@@ -1,0 +1,18 @@
+import type { MCState, MCAction } from '../../types';
+
+export function privilegeReducer(state: MCState, action: MCAction): MCState {
+    switch (action.type) {
+        case 'SET_PRIVILEGE_STATUS':
+            return {
+                ...state,
+                privileges: state.privileges.map(p =>
+                    p.id === action.cardId
+                        ? { ...p, status: action.status, suspendedUntil: action.suspendedUntil }
+                        : p,
+                ),
+            };
+
+        default:
+            return state;
+    }
+}
